generators out of Graph class
