+++ graphtool/graph/__init__.py
@@ -1,3 +1,4 @@
-from .graph import Graph
 from .vertex_edge import Vertex, Edge
+from .graph import Graph
 from .orientedGraph import OrientedGraph
+from .generator import GraphGenerator
