+++ tests/test_graph_generators.py
@@ -1,5 +1,6 @@
 import pytest
 from graphtool.graph import *
+from graphtool.graph.generator import *
 
 
 @pytest.fixture
@@ -9,18 +10,18 @@ def triangle():
 
 
 def test_cycle(triangle):
-    assert Graph.cycle(3) == triangle
+    assert GraphGenerator.cycle(3) == triangle
 
 
 def test_clique(triangle):
-    assert Graph.clique(3) == triangle
+    assert GraphGenerator.clique(3) == triangle
 
 
 def test_random_graph_generator():
-    assert Graph.erdos_renyi_proba(10, 0) == Graph.empty(10)
-    assert Graph.erdos_renyi_proba(10, 1) == Graph.clique(10)
-    graph1 = Graph.erdos_renyi_edge(100, 250)
-    graph2 = Graph.erdos_renyi_proba(100, 0.5)
+    assert GraphGenerator.erdos_renyi_proba(10, 0) == GraphGenerator.empty(10)
+    assert GraphGenerator.erdos_renyi_proba(10, 1) == GraphGenerator.clique(10)
+    graph1 = GraphGenerator.erdos_renyi_edge(100, 250)
+    graph2 = GraphGenerator.erdos_renyi_proba(100, 0.5)
     assert len(graph1.edges()) == 250
     assert len(graph1.vertices()) == 100
     assert len(graph2.vertices()) == 100
