+++ tests/test_graph_properties.py
@@ -17,7 +17,7 @@ def test_adjacency(triangle):
 def test_find_isolated_vertices():
     graph1 = Graph.from_edge_list("graph_examples/triangle_edge_list.txt")
     assert graph1.find_isolated_vertices() == []
-    graph2 = Graph.empty(2)
+    graph2 = GraphGenerator.empty(2)
     assert set(graph2.find_isolated_vertices()) == {Vertex(0), Vertex(1)}
 
 
@@ -29,13 +29,13 @@ def test_neighbours():
 def test_density():
     graph1 = Graph.from_edge_list("graph_examples/triangle_edge_list.txt")
     assert graph1.density() == 1
-    graph2 = Graph.empty(2)
+    graph2 = GraphGenerator.empty(2)
     assert graph2.density() == 0
 
 
 def test_diameter(triangle):
     assert diameter(triangle) == 1
-    assert diameter(Graph.empty(2)) == float("inf")
+    assert diameter(GraphGenerator.empty(2)) == float("inf")
 
 
 """
